Levantamiento observaciones SCTR al 19-07-2024

+++ MDS.Dto/MotivoDto.cs
@@ -2,9 +2,9 @@
 {
     public class MotivoDto
     {
-        public long id_motivo { get; set; }
-        public long tipo_atencion { get; set; }
-        public long tipo_pase { get; set; }
+        public int id_motivo { get; set; }
+        public int tipo_atencion { get; set; }
+        public int tipo_pase { get; set; }
         public string descripcion { get; set; }
         public Boolean skill { get; set; }
         //public int usuario_creacion { get; set; }
@@ -14,7 +14,7 @@
     }
     public class MotivoComboDto
     {
-        public long id_motivo { get; set; }
+        public int id_motivo { get; set; }
         public string descripcion { get; set; }
     }
 }
